wat-rtc-wkup: standby "works"
this now enters "standby" at ~0uA (I'd need to fiddle probes faster to
switch ranges) and then 5 seconds later "comes back" to 2.5mA, but
doesn't properly restart blinking.  Let's move on here, we've at least
got something working.

We're going to get a LPUART working so we can try and poke this when we
don't have the debugger.